fix reviews helper methods that were breaking the ability to see reviews if not logged in

+++ app/helpers/reviews_helper.rb
@@ -28,12 +28,12 @@ module ReviewsHelper
   end
 
   def delete_review(review)
-    if current_user.id == review.user_id
+    if current_user && current_user.id == review.user_id
       button_to "Delete Review", review_path(review), method: :delete, class: "btn"
     end
   end
   def edit_review(review)
-    if current_user.id == review.user_id
+    if current_user && current_user.id == review.user_id
       button_to "Edit Review", edit_review_path(review), method: :get, class: "btn"
     end
   end
